pyproject.toml
[tool.poetry]
name = "defect-dojo-api-v-2-client"
version = "2.25.1"
description = "A client library for accessing Defect Dojo API v2"

authors = []

readme = "README.md"
packages = [
    {include = "defect_dojo_api_v_2_client"},
]
include = ["CHANGELOG.md", "defect_dojo_api_v_2_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.20.0,<0.25.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
